move preset functions to separate file
