+++ packages/dep-check/src/profiles.ts
@@ -10,7 +10,6 @@ import { default as defaultPreset } from "./presets/microsoft/react-native";
 import type {
   MetaPackage,
   Package,
-  Preset,
   Profile,
   ProfileMap,
   ProfilesInfo,
@@ -183,71 +182,6 @@ export function profilesSatisfying(
 ): ProfileVersion[] {
   const versions = getProfileVersionsFor(versionOrRange);
   return profiles.filter((v) => versions.includes(v));
-}
-
-function v2_compileRequirements(
-  requirements: string[]
-): ((pkg: MetaPackage | Package) => boolean)[] {
-  const includePrerelease = { includePrerelease: true };
-  return requirements.map((req) => {
-    const [requiredPackage, requiredVersionRange] = req.split("@");
-    return (pkg: MetaPackage | Package) => {
-      return (
-        pkg.name === requiredPackage &&
-        "version" in pkg &&
-        semverSatisfies(
-          semverCoerce(pkg.version) || "0.0.1",
-          requiredVersionRange,
-          includePrerelease
-        )
-      );
-    };
-  });
-}
-
-function v2_loadPreset(preset: string, projectRoot: string): Preset {
-  try {
-    return require("./presets/" + preset).default;
-  } catch (_) {
-    return require(require.resolve(preset, { paths: [projectRoot] }));
-  }
-}
-
-export function v2_filterPreset(
-  requirements: string[],
-  preset: Preset
-): Preset {
-  const filteredPreset: Preset = {};
-  const reqs = v2_compileRequirements(requirements);
-  for (const [profileName, profile] of Object.entries(preset)) {
-    const packages = Object.values(profile);
-    const satisfiesRequirements = reqs.every((predicate) =>
-      packages.some(predicate)
-    );
-    if (satisfiesRequirements) {
-      filteredPreset[profileName] = profile;
-    }
-  }
-
-  return filteredPreset;
-}
-
-export function v2_mergePresets(
-  presets: string[],
-  projectRoot: string
-): Preset {
-  const mergedPreset: Preset = {};
-  for (const presetName of presets) {
-    const preset = v2_loadPreset(presetName, projectRoot);
-    for (const [profileName, profile] of Object.entries(preset)) {
-      mergedPreset[profileName] = {
-        ...mergedPreset[profileName],
-        ...profile,
-      };
-    }
-  }
-
-  return mergedPreset;
 }
 
 export function resolveCustomProfiles(
